HGA/requirements.txt
# requirements.txt

# Core Framework
crewai>=0.60.0
crewai-tools>=0.12.0

# LLM & API
openai>=1.3.0
tiktoken>=0.5.0

# Configuration & Data Handling
python-dotenv>=1.0.0
pydantic>=2.5.0
PyYAML>=6.0.1
jsonschema>=4.20.0

# Command-Line Interface
click>=8.1.0

# Knowledge Base & RAG (for future use)
qdrant-client>=1.7.0
sentence-transformers>=2.2.0

# Documentation (for building docs)
mkdocs>=1.5.0
mkdocs-material>=9.4.0
mkdocstrings[python]>=0.24.0

# Testing (install with `pip install -r requirements-test.txt`)
# pytest>=7.4.0
# pytest-mock>=3.12.0
# pytest-cov>=4.1.0
# requirements.txt
...
# Web API for Langflow Integration
fastapi>=0.104.0
uvicorn[standard]>=0.24.0